Housekeeping - Moved User Service Events Lambda
- Moved user service event lambda
- Output binaries go to the bin folder
